Trait tokio_io::codec::Encoder

pub trait Encoder {
    type Item;
    type Error: From<Error>;
    fn encode(&mut self,
              item: Self::Item,
              dst: &mut BytesMut)
              -> Result<(), Self::Error>;
}

Trait of helper objects to write out messages as bytes, for use with FramedWrite.

Associated Types

type Item

The type of items consumed by the Encoder

type Error: From<Error>

The type of encoding errors.

FramedWrite requires Encoders errors to implement From<io::Error> in the interest letting it return Errors directly.

Required Methods

fn encode(&mut self,
item: Self::Item,
dst: &mut BytesMut)
-> Result<(), Self::Error>

Encodes a frame into the buffer provided.

This method will encode msg into the byte buffer provided by buf. The buf provided is an internal buffer of the Framed instance and will be written out when possible.