Function tokio_io::io::lines

pub fn lines<A>(a: A) -> Lines<A> where
    A: AsyncRead + BufRead, 

Creates a new stream from the I/O object given representing the lines of input that are found on A.

This method takes an asynchronous I/O object, a, and returns a Stream of lines that the object contains. The returned stream will reach its end once a reaches EOF.