tokio_io::io

Struct ReadUntil

pub struct ReadUntil<A> { /* private fields */ }

A future which can be used to easily read the contents of a stream into a vector until the delimiter is reached.

Created by the read_until function.