Struct tokio_io::io::Window

pub struct Window<T> { /* fields omitted */ }

A owned window around an underlying buffer.

Normally slices work great for considering sub-portions of a buffer, but unfortunately a slice is a borrowed type in Rust which has an associated lifetime. When working with future and async I/O these lifetimes are not always appropriate, and are sometimes difficult to store in tasks. This type strives to fill this gap by providing an “owned slice” around an underlying buffer of bytes.

A Window<T> wraps an underlying buffer, T, and has configurable start/end indexes to alter the behavior of the AsRef<[u8]> implementation that this type carries.

This type can be particularly useful when working with the write_all combinator in this crate. Data can be sliced via Window, consumed by write_all, and then earned back once the write operation finishes through the into_inner method on this type.

Implementations

impl<T: AsRef<[u8]>> Window<T>

pub fn new(t: T) -> Window<T>

Creates a new window around the buffer t defaulting to the entire slice.

Further methods can be called on the returned Window<T> to alter the window into the data provided.

pub fn get_ref(&self) -> &T

Gets a shared reference to the underlying buffer inside of this Window.

pub fn get_mut(&mut self) -> &mut T

Gets a mutable reference to the underlying buffer inside of this Window.

pub fn into_inner(self) -> T

Consumes this Window, returning the underlying buffer.

pub fn start(&self) -> usize

Returns the starting index of this window into the underlying buffer T.

pub fn end(&self) -> usize

Returns the end index of this window into the underlying buffer T.

pub fn set_start(&mut self, start: usize) -> &mut Window<T>

Changes the starting index of this window to the index specified.

Returns the windows back to chain multiple calls to this method.

Panics

This method will panic if start is out of bounds for the underlying slice or if it comes after the end configured in this window.

pub fn set_end(&mut self, end: usize) -> &mut Window<T>

Changes the end index of this window to the index specified.

Returns the windows back to chain multiple calls to this method.

Panics

This method will panic if end is out of bounds for the underlying slice or if it comes before the start configured in this window.