Struct tokio_io::length_delimited::FramedWrite

pub struct FramedWrite<T, B: IntoBuf = BytesMut> {
    inner: T,
    builder: Builder,
    frame: Option<Chain<Cursor<BytesMut>, B::Buf>>,
}

Adapts a byte stream to a Sink accepting entire frame values.

See module level documentation for more detail.

Fields
inner: Tbuilder: Builderframe: Option<Chain<Cursor<BytesMut>, B::Buf>>
Methods

impl<T: AsyncWrite, B: IntoBuf> FramedWrite<T, B>

Important traits for FramedWrite<T, B>
impl<T: Read, B: IntoBuf> Read for FramedWrite<T, B>

pub fn new(inner: T) -> FramedWrite<T, B>

Creates a new FramedWrite with default configuration values.

impl<T, B: IntoBuf> FramedWrite<T, B>

pub fn max_frame_length(&self) -> usize

Returns the current max frame setting

This is the largest size this codec will write to the wire. Larger frames will be rejected.

pub fn set_max_frame_length(&mut self, val: usize)

Updates the max frame setting.

The change takes effect the next time a frame is encoded. In other words, if a frame is currently in process of being encoded with a frame size greater than val but less than the max frame length in effect before calling this function, then the frame will be allowed.

pub fn get_ref(&self) -> &T

Returns a reference to the underlying I/O stream wrapped by FramedWrite.

Note that care should be taken to not tamper with the underlying stream of data coming in as it may corrupt the stream of frames otherwise being worked with.

pub fn get_mut(&mut self) -> &mut T

Returns a mutable reference to the underlying I/O stream wrapped by FramedWrite.

Note that care should be taken to not tamper with the underlying stream of data coming in as it may corrupt the stream of frames otherwise being worked with.

pub fn into_inner(self) -> T

Consumes the FramedWrite, returning its underlying I/O stream.

Note that care should be taken to not tamper with the underlying stream of data coming in as it may corrupt the stream of frames otherwise being worked with.

impl<T: AsyncWrite, B: IntoBuf> FramedWrite<T, B>

fn do_write(&mut self) -> Poll<(), Error>

fn set_frame(&mut self, buf: B::Buf) -> Result<()>