Crate tokio_io

Core I/O traits and combinators when working with Tokio.

A description of the high-level I/O combinators can be found online in addition to a description of the low level details.

Modules

split
Split a single value implementing AsyncRead + AsyncWrite into separate AsyncRead and AsyncWrite handles.

Structs

BufReader
The BufReader struct adds buffering to any reader.

BufStream
Wraps a type that is [AsyncWrite] and [AsyncRead], and buffers its input and output.

BufWriter
Wraps a writer and buffers its output.

Traits

AsyncBufRead
Read bytes asynchronously.

AsyncBufReadExt
An extension trait which adds utility methods to AsyncBufRead types.

AsyncRead
Read bytes asynchronously.

AsyncReadExt
An extension trait which adds utility methods to AsyncRead types.

AsyncWrite
Writes bytes asynchronously.

AsyncWriteExt
An extension trait which adds utility methods to AsyncWrite types.

Buf
Read bytes from a buffer.

BufMut
A trait for values that provide sequential write access to bytes.