Crate tokio_io

Core I/O traits and combinators when working with Tokio.

Note: This crate has been deprecated in tokio 0.2.x and has been moved into tokio::io.

A description of the high-level I/O combinators can be found online in addition to a description of the low level details.

Modules

io

I/O conveniences when working with primitives in tokio-core

Macros

try_nb

A convenience macro for working with io::Result<T> from the Read and Write traits.

Traits

AsyncRead

Read bytes asynchronously.

AsyncWrite

Writes bytes asynchronously.

Type Definitions

IoFuture

A convenience typedef around a Future whose error component is io::Error

IoStream

A convenience typedef around a Stream whose error component is io::Error